+++ visual/tracks.h
@@ -35,10 +35,10 @@ public:
         {50, 100, 200}
     };
     bool active[128] = { [0 ... 127 ] = 1 };
-    short shape[128] = {0};
-    short track_blur[128] = {0};
-    short interconnect[128] = {0};
-    short colorScheme[128] = {0};
+    int shape[128] = {0};
+    int track_blur[128] = {0};
+    int interconnect[128] = {0};
+    int colorScheme[128] = {0};
     int max_connect_dist = 100; // toDo: transfer this to rendersetup and give the user the possibility of editing it there
     int maxBlur = 20;
